split PrettyYamlRenderer.scala
